+++ R/pluck.R
@@ -1,9 +1,16 @@
-#' Pluck or chuck a single element from a vector or environment
+#' Safely get or set an element deep within a nested data structure
 #'
-#' `pluck()` and `chuck()` implement a generalised form of `[[` that
-#' allow you to index deeply and flexibly into data structures.
-#' `pluck()` consistently returns `NULL` when an element does not
-#' exist while `chuck()` always throws (or chucks) an error.
+#' @description
+#' `pluck()` implements a generalised form of `[[` that allow you to index
+#' deeply and flexibly into data structures. It always succeeds, returning
+#' `.default` if the index you are trying to access does not exist or is `NULL`.
+#'
+#' `pluck<-()` is the assignment equivalent, allowing you to modify an object
+#' deped within a nested data structure.
+#'
+#' `pluck_exists()` tells you whether or not an object exists using the
+#' same rules as pluck (i.e. a `NULL` element is equivalent to an absent
+#' element).
 #'
 #' @param .x,x A vector or environment
 #' @param ... A list of accessors for indexing into the object. Can be
@@ -35,12 +42,12 @@
 #' @seealso [attr_getter()] for creating attribute getters suitable
 #'   for use with `pluck()` and `chuck()`. [modify_in()] for
 #'   applying a function to a pluck location.
+#' @export
 #' @examples
 #' # Let's create a list of data structures:
 #' obj1 <- list("a", list(1, elt = "foo"))
 #' obj2 <- list("b", list(2, elt = "bar"))
 #' x <- list(obj1, obj2)
-#'
 #'
 #' # pluck() provides a way of retrieving objects from such data
 #' # structures using a combination of numeric positions, vector or
@@ -69,43 +76,13 @@
 #' # You can also supply a default value for non-existing elements:
 #' pluck(x, 10, .default = NA)
 #'
-#' # If you prefer to consistently fail for non-existing elements, use
-#' # the opinionated variant chuck():
-#' chuck(x, 1)
-#' try(chuck(x, 10))
-#' try(chuck(x, 1, 10))
-#'
-#'
 #' # The map() functions use pluck() by default to retrieve multiple
 #' # values from a list:
-#' map(x, 2)
-#'
-#' # Pass multiple indexes with a list:
-#' map(x, list(2, "elt"))
-#'
-#' # This is equivalent to:
-#' map(x, pluck, 2, "elt")
-#'
-#' # You can also supply a default:
-#' map(x, list(2, "elt", 10), .default = "superb default")
-#'
-#' # Or use the strict variant:
-#' try(map(x, chuck, 2, "elt", 10))
-#'
-#'
-#' # You can also assign a value in a pluck location with pluck<-:
-#' pluck(x, 2, 2, "elt") <- "quuux"
-#' x
-#'
-#' # This is a shortcut for the prefix function assign_in():
-#' y <- assign_in(x, list(2, 2, "elt"), value = "QUUUX")
-#' y
-#'
+#' map_chr(x, 1)
+#' map_int(x, c(2, 1))
 #'
 #' # pluck() also supports accessor functions:
 #' my_element <- function(x) x[[2]]$elt
-#'
-#' # The accessor can then be passed to pluck:
 #' pluck(x, 1, my_element)
 #' pluck(x, 2, my_element)
 #'
@@ -115,14 +92,27 @@
 #' # expression:
 #' my_element(x[[1]])
 #'
-#'
 #' # If you have a list of accessors, you can splice those in with `!!!`:
 #' idx <- list(1, my_element)
 #' pluck(x, !!!idx)
-#' @export
 pluck <- function(.x, ..., .default = NULL) {
   check_dots_unnamed()
   pluck_raw(.x, list2(...), .default = .default)
+}
+
+#' @rdname pluck
+#' @inheritParams modify_in
+#' @export
+`pluck<-` <- function(.x, ..., value) {
+  assign_in(.x, list2(...), value)
+}
+
+#' @rdname pluck
+#' @export
+pluck_exists <- function(.x, ...) {
+  check_dots_unnamed()
+
+  !is_zap(pluck_raw(.x, list2(...), .default = zap()))
 }
 
 pluck_raw <- function(.x, index, .default = NULL) {
@@ -135,8 +125,27 @@ pluck_raw <- function(.x, index, .default = NULL) {
   )
 }
 
-#' @rdname pluck
+#' Get an element deep within a nested data structure, failing if it doesn't
+#' exist
+#'
+#' `chuck()` implements a generalised form of `[[` that allow you to index
+#' deeply and flexibly into data structures. If the index you are trying to
+#' access does not exist (or is `NULL`), it will throw (i.e. chuck) an error.
+#'
+#' @seealso [pluck()] for a quiet equivalent.
+#' @inheritParams pluck
 #' @export
+#' @examples
+#' x <- list(a = 1, b = 2)
+#'
+#' # When indexing an element that doesn't exist `[[` sometimes returns NULL:
+#' x[["y"]]
+#' # and sometimes errors:
+#' try(x[[3]])
+#'
+#' # chuck() consistently errors:
+#' try(chuck(x, "y"))
+#' try(chuck(x, 3))
 chuck <- function(.x, ...) {
   check_dots_unnamed()
 
@@ -147,13 +156,6 @@ chuck <- function(.x, ...) {
     missing = NULL,
     strict = TRUE
   )
-}
-
-#' @rdname pluck
-#' @inheritParams modify_in
-#' @export
-`pluck<-` <- function(.x, ..., value) {
-  assign_in(.x, list2(...), value)
 }
 
 reduce_subset_call <- function(init, idx) {
@@ -210,4 +212,76 @@ is_index <- function(x) {
 attr_getter <- function(attr) {
   force(attr)
   function(x) attr(x, attr, exact = TRUE)
+}
+
+
+#' Modify a pluck location
+#'
+#' @description
+#'
+#' * `assign_in()` takes a data structure and a [pluck] location,
+#'   assigns a value there, and returns the modified data structure.
+#'
+#' * `modify_in()` applies a function to a pluck location, assigns the
+#'   result back to that location with [assign_in()], and returns the
+#'   modified data structure.
+#'
+#' @inheritParams pluck
+#' @param .f A function to apply at the pluck location given by `.where`.
+#' @param ... Arguments passed to `.f`.
+#' @param .where,where A pluck location, as a numeric vector of
+#'   positions, a character vector of names, or a list combining both.
+#'   The location must exist in the data structure.
+#' @seealso [pluck()]
+#' @export
+#' @examples
+#' # Recall that pluck() returns a component of a data structure that
+#' # might be arbitrarily deep
+#' x <- list(list(bar = 1, foo = 2))
+#' pluck(x, 1, "foo")
+#'
+#' # Use assign_in() to modify the pluck location:
+#' str(assign_in(x, list(1, "foo"), 100))
+#' # Or zap to remove it
+#' str(assign_in(x, list(1, "foo"), zap()))
+#'
+#' # Like pluck(), this works even when the element (or its parents) don't exist
+#' pluck(x, 1, "baz")
+#' str(assign_in(x, list(2, "baz"), 100))
+#'
+#' # modify_in() applies a function to that location and update the
+#' # element in place:
+#' modify_in(x, list(1, "foo"), ~ .x * 200)
+#'
+#' # Additional arguments are passed to the function in the ordinary way:
+#' modify_in(x, list(1, "foo"), `+`, 100)
+modify_in <- function(.x, .where, .f, ...) {
+  .where <- as.list(.where)
+  .f <- rlang::as_function(.f)
+
+  value <- .f(pluck(.x, !!!.where), ...)
+  assign_in(.x, .where, value)
+}
+#' @rdname modify_in
+#' @param value A value to replace in `.x` at the pluck location.
+#'   Use `zap()` to instead remove the element.
+#' @export
+assign_in <- function(x, where, value) {
+  n <- length(where)
+  if (n == 0) {
+    abort("`where` must contain at least one element")
+  } else if (n > 1) {
+    old <- pluck(x, where[[1]], .default = list())
+    if (!is_zap(value) || !identical(old, list())) {
+      value <- assign_in(old, where[-1], value)
+    }
+  }
+
+  if (is_zap(value)) {
+    x[[where[[1]]]] <- NULL
+  } else {
+    list_slice2(x, where[[1]]) <- value
+  }
+
+  x
 }