+++ tests/testthat/test-map-df.R
@@ -1,0 +1,51 @@
+test_that("_df/_dfc/_dfr are deprecated", {
+  expect_snapshot({
+    . <- map_df(list(), identity)
+    . <- map_dfr(list(), identity)
+    . <- map_dfc(list(), identity)
+
+    . <- map2_df(list(), list(), identity)
+    . <- map2_dfr(list(), list(), identity)
+    . <- map2_dfc(list(), list(), identity)
+
+    . <- imap_dfr(list(), identity)
+    . <- imap_dfc(list(), identity)
+
+    . <- pmap_df(list(), identity)
+    . <- pmap_dfr(list(), identity)
+    . <- pmap_dfc(list(), identity)
+  })
+})
+
+test_that("row and column binding work", {
+  local_options(lifecycle_verbosity = "quiet")
+
+  skip_if_not_installed("dplyr")
+  local_name_repair_quiet()
+
+  mtcar_mod <- mtcars %>%
+    split(.$cyl) %>%
+    map(~ lm(mpg ~ wt, data = .x))
+
+  f_coef <- function(x) as.data.frame(t(as.matrix(coef(x))))
+  expect_length(mtcar_mod %>% map_dfr(f_coef), 2)
+  expect_length(mtcar_mod %>% map_dfc(f_coef), 6)
+})
+
+test_that("data frame imap works", {
+  local_options(lifecycle_verbosity = "quiet")
+  skip_if_not_installed("dplyr")
+  x <- set_names(1:3)
+  expect_identical(imap_dfc(x, paste), imap_dfr(x, paste))
+})
+
+test_that("outputs are suffixes have correct type for data frames", {
+  local_options(lifecycle_verbosity = "quiet")
+  skip_if_not_installed("dplyr")
+  local_name_repair_quiet()
+
+  local_options(rlang_message_verbosity = "quiet")
+  x <- 1:3
+  expect_s3_class(pmap_dfr(list(x), as.data.frame), "data.frame")
+  expect_s3_class(pmap_dfc(list(x), as.data.frame), "data.frame")
+})